backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.29
GeoAlchemy2==0.14.7
pydantic==2.6.4
shapely==2.0.3
redis==5.0.3
asyncpg==0.29.0
python-dotenv==1.0.1
websockets==12.0
